Addon list saving and loading
Quite a bit of work went into this. Addons are saved in plaintext compared to the main config. This can be useful if you have any clientside mods such as ManiaHud or any palette altering wads.
Maybe in the future, they can use relative paths, that way, they can theoretically be "shared" in some manner...